+++ fractal/components/_preview-layouts/_recommendation-score-summary-preview.jsx
@@ -11,9 +11,7 @@ const page = ({ rendered_content, path }) => {
             <link media="all" rel="stylesheet" href={`http://localhost:3050/static/css/links/Link/Link.css`}/>
             <link media="all" rel="stylesheet" href={`http://localhost:3050/static/css/cards/Card/Card.css`}/>
             <link media="all" rel="stylesheet" href={`http://localhost:3050/static/css/loading-spinners/LoadingSpinner/LoadingSpinner.css`}/>
-            <link media="all" rel="stylesheet" href={`http://localhost:3050/static/css/assessments/RecommendationSummary/RecommendationSummary.css`}/>
             <link media="all" rel="stylesheet" href={`http://localhost:3050/static/css/assessments/RecommendationScoreSummary/RecommendationScoreSummary.css`}/>
-            <link media="all" rel="stylesheet" href={`http://localhost:3050/static/css/assessments/RecommendationSummaryGroup/RecommendationSummaryGroup.css`}/>
             <style>{"\
                 .base-preview{\
                   margin: 5px;\
@@ -21,7 +19,7 @@ const page = ({ rendered_content, path }) => {
               "}</style>
             <div className="base-preview" dangerouslySetInnerHTML={{__html: rendered_content}}/>
             <script src={`http://localhost:3050/static/vendors/snap.svg-min.js`}/>
-            <script src={`http://localhost:3050/static/js/RecommendationSummaryGroup.js`}/>
+            <script src={`http://localhost:3050/static/js/RecommendationScoreSummary.js`}/>
         </div>
     );
 };