+++ fractal/components/assessments/RecommendationScoreSummary/index.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+
+import RecommendationScoreSummary from './RecommendationScoreSummary.jsx';
+import config from './RecommendationScoreSummary.config';
+
+ReactDOM.render(
+    <RecommendationScoreSummary {...config.context} />,
+    document.getElementsByClassName('base-preview')[0]
+);