+++ gulp/build-configs/webpack.config.style-guide.js
@@ -8,7 +8,8 @@ module.exports = {
         StickyBar: './fractal/components/stickies/StickyBar/index.jsx',
         SVGArcChart: './fractal/components/charts/SVGArcChart/index.jsx',
         RecommendationChart: './fractal/components/assessments/RecommendationChart/index.jsx',
-        RecommendationSummaryGroup: './fractal/components/assessments/RecommendationSummaryGroup/index.jsx'
+        RecommendationSummaryGroup: './fractal/components/assessments/RecommendationSummaryGroup/index.jsx',
+        RecommendationScoreSummary: './fractal/components/assessments/RecommendationScoreSummary/index.jsx'
     },
     output: {
         path: path.resolve(__dirname, '../../_site/js'),